Change to catch only TX:12345-... variables
Error messages misleading with TX.0 in it when it should be TX:00000-..
for example
Expanded "TX:/^\d/" to "TX:0|TX:981245-Detects basic SQL authentication bypass attempts 2/3-OWASP_CRS/WEB_ATTACK/SQLI-ARGS:TESTARGS"
Then TX:0 matches. If you want this information for further processing it is not useful to get TX:0 because it has no information which rule matched.