fix invalid skipAfter when client sends no Accept header
Rule 920300 (now in PL2) performs a skipAfter:END_ACCEPT_CHECK.
However, the END_ACCEPT_CHECK marker has remained at the old
location, above it. So, if this rule fires, we have a skipAfter loop.

This causes the bug that if paranoia level is 2 or higher, and the
client sends no Accept header, the rules and files after rule 920300
are no longer processed.

This is fixed by removing skipAfter and the marker, while keeping
the "empty Accept header" check at the old paranoia level.