setup.conf: retire tx.anomaly_score_blocking (rule 900004)
It is no longer needed for the user to configure this variable.
The variable was only evaluated in rules 949110 and 959100.
However, in self-contained mode, these rules are a no-op, because
in case of a rule match, the processing has already stopped, and
in case of no match, the scores should be 0.