fix: change msg values to not span multiple lines
For `SecRule` `msg`, the value is some format string.  It may be
tempting to wrap it across multiple lines, but that does not seem to be
possible as trailing backslashes don't escape newlines.  For instance,
having

```
    msg:'Blah blah
blah',\
```

causes the message to appear in logs as

```
Blah blah\x0ablah
```

and having

```
    msg:'Blah blah\
blah',\
```

causes the message to appear in logs as

```
Blah blah\\x0ablah
```

Fix all instances of a `msg` value wrapping across newlines: just one in
`REQUEST-901-INITIALIZATION.conf`.